Loading from db

+++ desktop/gui/src-tauri/migrations/2024-08-28-055132_create_file_blocks/up.sql
@@ -5,7 +5,7 @@ CREATE TABLE file_blocks(
     file_id VARCHAR(50) NOT NULL,
     -- index of the block within the file
     -- the size of the block is determined by its index and the file's chunk size
-    block_index INTEGER NOT NULL,
+    block_index BIGINT NOT NULL,
     status VARCHAR(50) NOT NULL,
     PRIMARY KEY (id, file_id),
     FOREIGN KEY (file_id) REFERENCES files(id)
